+++ egui_commonmark/examples/macro.rs
@@ -1,8 +1,8 @@
 //! Make sure to run this example from the repo directory and not the example
 //! directory. To see all the features in full effect, run this example with
-//! `cargo r --example macro --features macro,better_syntax_highlighting`
+//! `cargo r --example macro --features macros,better_syntax_highlighting`
 //! Add `light` or `dark` to the end of the command to specify theme. Default
-//! is light. `cargo r --example macro --features macro,better_syntax_highlighting -- dark`
+//! is light. `cargo r --example macro --features macros,better_syntax_highlighting -- dark`
 
 use eframe::egui;
 use egui_commonmark::*;